+++ WoWsShipBuilder.UI/Translations/Translation.Designer.cs
@@ -1239,6 +1239,15 @@ namespace WoWsShipBuilder.UI.Translations {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Air Drag.
+        /// </summary>
+        public static string ShipStats_AirDrag {
+            get {
+                return ResourceManager.GetString("ShipStats_AirDrag", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Airstrike.
         /// </summary>
         public static string ShipStats_Airstrike {
@@ -1298,6 +1307,18 @@ namespace WoWsShipBuilder.UI.Translations {
         public static string ShipStats_Backward {
             get {
                 return ResourceManager.GetString("ShipStats_Backward", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to If &quot;module blast protection&quot; + &quot;armament blast penetration&quot; &gt; 0 
+        ///     The module can be incapacitated.
+        ///If &quot;module blast protection&quot; + &quot;armament blast penetration&quot; &lt;= 0
+        ///     The module can not be incapacitated..
+        /// </summary>
+        public static string ShipStats_BlastExplanation {
+            get {
+                return ResourceManager.GetString("ShipStats_BlastExplanation", resourceCulture);
             }
         }
         
@@ -1509,7 +1530,7 @@ namespace WoWsShipBuilder.UI.Translations {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Formula at range X.
+        ///   Looks up a localized string similar to Formula at range X(Km).
         /// </summary>
         public static string ShipStats_DispFormulaTooltip {
             get {
@@ -1536,6 +1557,24 @@ namespace WoWsShipBuilder.UI.Translations {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Engine blast protection.
+        /// </summary>
+        public static string ShipStats_EngineBlastProtection {
+            get {
+                return ResourceManager.GetString("ShipStats_EngineBlastProtection", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Explosion radius.
+        /// </summary>
+        public static string ShipStats_ExplosionRadius {
+            get {
+                return ResourceManager.GetString("ShipStats_ExplosionRadius", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Max fires.
         /// </summary>
         public static string ShipStats_FireAmount {
@@ -1550,6 +1589,15 @@ namespace WoWsShipBuilder.UI.Translations {
         public static string ShipStats_FireChance {
             get {
                 return ResourceManager.GetString("ShipStats_FireChance", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Fire chance per salvo.
+        /// </summary>
+        public static string ShipStats_FireChancePerSalvo {
+            get {
+                return ResourceManager.GetString("ShipStats_FireChancePerSalvo", resourceCulture);
             }
         }
         
@@ -1851,6 +1899,15 @@ namespace WoWsShipBuilder.UI.Translations {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Mass.
+        /// </summary>
+        public static string ShipStats_Mass {
+            get {
+                return ResourceManager.GetString("ShipStats_Mass", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Max number of fires.
         /// </summary>
         public static string ShipStats_MaxFireAmount {
@@ -2148,6 +2205,15 @@ namespace WoWsShipBuilder.UI.Translations {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Rudder blast protection.
+        /// </summary>
+        public static string ShipStats_RudderBlastProtection {
+            get {
+                return ResourceManager.GetString("ShipStats_RudderBlastProtection", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Second ping duration.
         /// </summary>
         public static string ShipStats_SecondPingDuration {
@@ -2247,6 +2313,15 @@ namespace WoWsShipBuilder.UI.Translations {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Blast penetration.
+        /// </summary>
+        public static string ShipStats_SplashCoeff {
+            get {
+                return ResourceManager.GetString("ShipStats_SplashCoeff", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Starting planes on deck.
         /// </summary>
         public static string ShipStats_StartingNumberOnDeck {
@@ -2270,6 +2345,15 @@ namespace WoWsShipBuilder.UI.Translations {
         public static string ShipStats_TheoreticalDPM {
             get {
                 return ResourceManager.GetString("ShipStats_TheoreticalDPM", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Theoretical FPM .
+        /// </summary>
+        public static string ShipStats_TheoreticalFPM {
+            get {
+                return ResourceManager.GetString("ShipStats_TheoreticalFPM", resourceCulture);
             }
         }