+++ ReleaseNotes.md
@@ -5,9 +5,13 @@
 - Add dispersion plot in the dispersion and ballistics window
 - Add optional telemetry data in settings. Telemetry data is disabled by default and needs to be enabled manually
 - Add a tooltip for dispersion formula in main battery column (#10)
+- Add several parameters to ships armaments
+- Add blast protection parameter for engine and rudder (#9)
+- Add torpedo layout (#50)
 
 ## Changes
 - Windows now automatically resize to fit on smaller screens or on screens with dpi scaling enabled (#56)
+- Dispersion formulas are now in Km istead of m
 
 ## Bugfixes
 - Show missing modules on aircraft carriers (#38)
@@ -23,6 +27,9 @@
 - Fixed engine upgrade modifier calculation
 - Fixed bug in incremental image updates using the wrong version name
 - Fixed wrong unit for RoF in ship stats
+- Fixed incorrect value for theoretical DPM
+- Fixed incorrect guns name display
+- Fixed overmatch value not being displayed or being shown incorrectly 
 ___
 To download the program for the first time, you need to download only the Setup.exe file. The other files are used by the application for automatic updates.